login-jwt/backend/requirements.txt
flask
pyjwt
flask-cors
werkzeug
bcrypt
